now that we have to use C99, use long long in flag_str struct

+++ src/dfc.c
@@ -789,7 +789,7 @@ disp(struct list *lst, char *fstfilter, char *fsnfilter, struct display *disp)
  * All the flags found in *BSD and Mac OS X, alphabetically sorted.
  */
 struct flag_str {
-	long flag;
+	long long flag;
 	const char *str;
 } possible_flags[] = {
 #if defined(__FreeBSD__)
